% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldClimData-package.R
\docType{package}
\name{WorldClimData-package}
\alias{WorldClimData}
\alias{WorldClimData-package}
\title{WorldClimData: Download WorldClim Data}
\description{
A faster way to download WorldClim data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brunomioto/WorldClimData}
  \item \url{https://brunomioto.github.io/WorldClimData/}
  \item Report bugs at \url{https://github.com/brunomioto/WorldClimData/issues}
}

}
\author{
\strong{Maintainer}: Bruno Mioto \email{brunomioto97@gmail.com}

}
\keyword{internal}
