% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldClimData-package.R
\docType{package}
\name{WorldClimData-package}
\alias{WorldClimData}
\alias{WorldClimData-package}
\title{WorldClimData: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brunomioto/WorldClimData}
  \item Report bugs at \url{https://github.com/brunomioto/WorldClimData/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
